+++ Int32ToIP/app/src/test/java/org/contest/IntegerToIPTest.java
@@ -40,7 +40,7 @@ class IntegerToIPTest {
 		return Integer.parseUnsignedInt(bits.toString(), 2);
 	}
 
-	private static String a(int x) {
+	private static String a(long x) {
 		StringBuilder b = new StringBuilder();
 		int i = 0;
